Rename nvim config directory
