+++ roles/neovim/files/init.lua
@@ -1,1 +1,1 @@
-require("edvin")
+require("config")
