+++ roles/neovim/files/lua/config/init.lua
@@ -1,9 +1,9 @@
-require("edvin.remap")
-require("edvin.set")
-require("edvin.packer")
+require("config.remap")
+require("config.set")
+require("config.packer")
 
 local augroup = vim.api.nvim_create_augroup
-local EdvinGroup = augroup('Edvin', {})
+local ConfGroup = augroup('Conf', {})
 
 local autocmd = vim.api.nvim_create_autocmd
 local yank_group = augroup('HighlightYank', {})
@@ -24,7 +24,7 @@ autocmd('TextYankPost', {
 })
 
 autocmd({"BufWritePre"}, {
-    group = EdvinGroup,
+    group = ConfGroup,
     pattern = "*",
     command = [[%s/\s\+$//e]],
 })
